refactor: rename typeUtils folder structure
