+++ jest.config.ts
@@ -34,7 +34,12 @@ const config: Config = {
 	coverageDirectory: 'coverage',
 
 	// An array of regexp pattern strings used to skip coverage collection
-	coveragePathIgnorePatterns: ['/node_modules/', '/src/testUtils/', '/src/test-utils.tsx'],
+	coveragePathIgnorePatterns: [
+		'/node_modules/',
+		'/src/testUtils/',
+		'/src/test-utils.tsx',
+		'/src/types'
+	],
 
 	// A list of reporter names that Jest uses when writing coverage reports
 	coverageReporters: [
