+++ sonar-project.properties
@@ -4,7 +4,7 @@
 # SPDX-License-Identifier: AGPL-3.0-only
 #
 sonar.sources=src/
-sonar.javascript.exclusions=**/*.test.*,**/typeUtils/*,**/test*.ts*,.storybook,**/*.stories.*
+sonar.javascript.exclusions=**/*.test.*,**/test*.ts*,.storybook,**/*.stories.*
 sonar.tests=src/
-sonar.test.inclusions=**/*.test.*,**/typeUtils/*,**/test*.ts*,**/*.stories.*
+sonar.test.inclusions=**/*.test.*,**/test*.ts*,**/*.stories.*
 sonar.cpd.exclusions=**/*.test.*
